apps/mail-search/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
psycopg2-binary>=2.9.0
anthropic>=0.40.0
boto3>=1.35.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
